Stage fixes from reading, fused matrix/spline math, and the AI-hash probes
Stages: Poke Floats' 27 float descs come from the DOL instead of an overlay
past the decompiled table; Mute City's parameter block is converted per
field (two colour-overlay scripts, two touch-line hit descriptors, the car
scalars) and its cars' spline joints are converted (JOBJ_SPLINE payloads
were dropped; Kongo Jungle and Yoshi's Story carry one each too); Final
Destination's, Venom's and Peach's Castle's colour-overlay script pointers
are resolved from file offsets (grMaterial_801C9604 takes a pointer-sized
argument now); Castle's layout gets its last 16 bytes; the touch-line
result is read as the hit descriptor it is. A class of truncated GObj/joint
pointers stored through u32 slots and read back as pointers is made
pointer-sized in Castle, Venom, Brinstar and Mute City, and Castle's
GroundVars12 overlay no longer reads half-pointers. The texture-desc
sanity guard no longer rejects this binary's own static data (Poke Floats'
toon texture). Two implicit declarations that truncated a returned pointer
to 32 bits are gone (grcastle's archive lookup, grvenom's resolver).

Math: HSD_MtxInverse, HSD_MtxInverseConcat, HSD_MtxInverseTranspose and
HSD_MtxScaledAdd follow the console's fused sequences (they were plain
arithmetic: the hurtbox inverse, robj constraints, skinning); every branch
of splGetSplinePoint and the arc-length integrator likewise; the
animation-rewind fmod is the console's divide-truncate-fnmsubs.

Probes: MELEE_DIDBG (CPU knockback-DI operands), MELEE_CPUCMDAT (who wrote
each CPU command). Both sides of the Samus/PStadium and Peach/Izumi AI-hash
divergences are now attributed: a 1-ulp kb_vel.x, and uninitialised stick
bytes the console fills from a leftover r5 and the fighter's own address.

+++ src/melee/gr/grzebes.c
@@ -361,9 +361,9 @@ void grZebes_801D8644(HSD_GObj* gobj)
     mat_gobj2 =
         grMaterial_801C8D44(0, 0, gp, &pos, 0, NULL, fn_801DAC90, NULL);
     grMaterial_801C8E08(mat_gobj2);
-    gp->gv.zebes5.x100 = (u32) mat_gobj2;
+    gp->gv.zebes5.x100 = (uintptr_t) mat_gobj2;
     grZebes_801DC9DC((s32) gobj);
-    gp->gv.zebes5.xFC = (u32) grZakoGenerator_801CA394(
+    gp->gv.zebes5.xFC = (uintptr_t) grZakoGenerator_801CA394(
         (UNK_T) &grZe_803E1B90, 0xA, (UNK_T) grZebes_801DCB64, 1.0f);
     mpJointSetB10(0);
     Ground_801C2FE0((Ground_GObj*) gobj);
@@ -679,7 +679,7 @@ void grZebes_801D9100(HSD_GObj* gobj)
     gp->gv.zebes4.xE0 = PC_GVPTR(mat_gobj);
     gp->gv.zebes4.xE4 = new_var;
     gp->gv.zebes4.xE8 = 0;
-    gp->gv.zebes4.xEC = (u32) grZakoGenerator_801CA394(
+    gp->gv.zebes4.xEC = (uintptr_t) grZakoGenerator_801CA394(
         (void*) &grZe_803E1C80, 4, (void*) grZebes_801DCBB0, 1.0f);
     Ground_801C2FE0(new_var2);
 }
@@ -764,7 +764,7 @@ void grZebes_801D9508(Ground_GObj* gobj)
     HSD_GObj* map_a_gobj = Ground_801C2BA4(6);
     PAD_STACK(0x10);
     HSD_ASSERT(909, map_a_gobj);
-    gp->gv.zebes.x4 = (u32) Ground_801C3FA4(map_a_gobj, 14);
+    gp->gv.zebes.x4 = (uintptr_t) Ground_801C3FA4(map_a_gobj, 14);
     gp->gv.zebes.x8 = 1;
     gp->gv.zebes.xA = (s16) (HSD_Randi(600) + 3000);
     gp->gv.zebes.x0_b0 = true;
